Add files via upload
Software for a PIC18F46K22 (or PIC18LF46K22) to monitor a tipping bucket rain gauge.  The code drives an RFM95W LoRa module to upload the rain tip total every 2 minutes or when there is a new tip.  It also sends battery voltage and local temperature.  Message is sent as 50 bytes.  See my weather station details on https://wordpress.com/home/andypageelectronics.wordpress.com
Schematic of the electronics is included in the files.